Move track initialization launchers into separate files
